вынес валидацию STDIN и ARGV в отдельный класс
