+++ .rubocop_todo.yml
@@ -1,10 +1,16 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2020-04-06 14:30:55 +0300 using RuboCop version 0.79.0.
+# on 2020-04-06 15:33:34 +0300 using RuboCop version 0.79.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
 # versions of RuboCop, may require this file to be generated again.
+
+# Offense count: 2
+# Configuration parameters: CountComments, ExcludedMethods.
+# ExcludedMethods: refine
+Metrics/BlockLength:
+  Max: 58
 
 # Offense count: 4
 RSpec/ExpectOutput:
@@ -17,7 +23,7 @@ Style/AsciiComments:
   Exclude:
     - 'spec/spec_helper.rb'
 
-# Offense count: 8
+# Offense count: 12
 # Cop supports --auto-correct.
 # Configuration parameters: AutoCorrect, AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, IgnoredPatterns.
 # URISchemes: http, https
