+++ exe/nmax
@@ -1,7 +1,7 @@
 #!/usr/bin/env ruby
 # frozen_string_literal: true
 
-require 'nmax/cli'
+require 'nmax'
 
 cli = Nmax::CLI.new
 cli.run
