+++ lib/nmax.rb
@@ -2,6 +2,7 @@
 
 require 'nmax/version'
 require 'nmax/cli'
+require 'nmax/cli/input_parser'
 
 module Nmax
   class Error < StandardError; end
