+++ lib/nmax/cli.rb
@@ -3,9 +3,14 @@
 module Nmax
   # This class is responsible for command line interface logic
   class CLI
+    attr_reader :input_parser
+
+    def initialize(input_parser = Nmax::CLI::InputParser.new)
+      @input_parser = input_parser
+    end
+
     def run
-      validate_argv
-      validate_stdin
+      input_parser.validate
 
       text = STDIN.read
 
@@ -17,20 +22,7 @@ module Nmax
     private
 
     def n_arg
-      @n_arg ||= ARGV[0].to_i
-    end
-
-    def validate_argv
-      abort 'В скрипт должен быть передан аргумент N.' if ARGV.empty?
-      abort 'Скрипт поддерживает передачу только одного аргумента N.' if ARGV.size > 1
-      abort 'Аргумент N должен быть числом, больше нуля.' if n_arg <= 0
-    end
-
-    def validate_stdin
-      return unless STDIN.tty?
-
-      abort 'Входящий поток не содержит текстовых данных, пример использования скрипта:'\
-            ' `cat sample_data_40GB.txt | nmax 10000`'
+      @n_arg ||= input_parser.n_arg
     end
   end
 end
